Fix child popup palette inheritance
Qt Quick Controls Popups are always top-level popups, even if declared
as a child of another popup. Therefore, e.g. a font set on a parent
popup should not propagate to a child popup. The same behavior applies
to palettes, but we regressed with this in
1875ad7f92cad270cc5857d71096a4b46c27c562 and/or
31f5c21ddb571d744efd1885c0687816b3a12225.

This patch fixes the regression by returning the window's palette in
QQuickPaletteProviderPrivateBase<I, Impl>::parentPalette() if I is
a QQuickPopup.

Fixes: QTBUG-115707
Change-Id: Ib89852787dd45e3edc9e846b6c182340c052f0fd
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 950cd5888e3098bc4c04fc0af4beabe265c78a6e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>