qmltyperegistrar: fix minor versions
If QML_ADDED_IN_MINOR_VERSION or QML_REMOVED_IN_MINOR_VERSION is used,
the QTypeRevision of the type will lack a major version. When we later
try to read them, that results in a value of 255. Instead of writing
out that value, fetch the value from the module, just like the engine
does at runtime.

Do the same for QML_REMOVED_IN_MINOR_VERSION.

Change-Id: I03d4364e1b62dda171dcf5bd4c327b1a00f616f5
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>