Clarify visibility behavior of Windows with transient parents
In ccbf0560b36515630328be44a20f4fe62eaec531 we started treating nested
Windows as being transient children of the containing Window.

Unlike QWindowPrivate::setVisible() for true child windows, we do
not have any logic to defer the creation or visibility of transient
children until their transient parent is visible. It's up to the
platform plugin to deal with this, which is not ideal.

With a declarative API, it was easy to get into this situation, as
the engine completes child components before parents, so a transient
child window with visible set to true, would be shown before its
parent.

In c831a8e85afee6b257f8b3b1c5d86e4375b41e8c we started deferring the
visibility of transient children until their transient parent was shown.

Finally, in 02c98b7590b060430dc5ce7e5821fa13087d25f7 a dedicated
transientParent property was added to Window, to allow the user
to override the transient parent. As part of this change, documentation
was added that described the deferred visibility for transient
children, but did so in a way that could lead users to believe
that it only applied to the automatic transient parents,
when in fact it applies to all transient parents, regardless
of how they were set.

The documentation now clarifies this by describing the two
behaviors independently. We also mention that the automatic
transient parent applies for Qt.createComponent and
Qt.createQmlObject as well, as it's nowadays the case.

Change-Id: I98f50696b2c4a79540fd223353980699af858817
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 0f3464cda31390c5e3ecef7e9c91d74a6c0ad663)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>