Fix memory leak in qquickpixmapcache
QtPdf depends on QQuickPixmap::loadImageFromDevice() to pass a QPdfFile
(which carries a pointer to the QPdfDocument) down into QPdfIOHandler
so that it can avoid needing to reopen the document. But the
introduction of loadImageFromDevice(), and then using the pixmap cache
so intensively to store a lot of rendered PDF pages, revealed a drastic
memory leak: many QQuickPixmapData instances remained in
QQuickPixmapStore::m_cache long after their refCounts became 0.
Quick scrolling through a PdfMultiPageView (especially if the zoom level
is also being changed dynamically) can easily generate a lot of
page-rendering jobs, some of which get cancelled early because the view
has moved on to other pages and/or zoom levels.

Here are several related fixes:

- QQuickPixmapReader::getImage() creates a "job" represented by
  the QQuickPixmapReply that is returned, but now we start the job
  in a separate method, to ensure that
  QQuickPixmap::loadImageFromDevice() succeeds in connecting the
  QQuickPixmapReply::destroyed() signal before the job completes
  and deletes itself.

- Use a QMutex to prevent QQuickPixmapStore::m_cache reentrancy:
  Rarely, a crash was possible when QQuickPixmapStore's QHash m_cache
  was modified between QHash::find() and dereferencing the iterator it
  returns, e.g. in QQuickPixmap::loadImageFromDevice().

- QQuickPixmapData::release() is needed in a few more places.

- QQuickPixmapReply::finished isn't emitted if the job is cancelled, so
  the lambda that called oldD->release() wasn't called, and that was the
  remaining leak. We tried connecting QObject::destroyed() earlier, but
  it was causing a crash because the mutex to protect m_cache from
  simultaneous modification and QHash::find() was not in place. Now it
  seems ok.

- Show "Number of leaked pixmaps: 0" in the qt.scenegraph.leaks logging
  category to make verification less ambiguous.

Pick-to: 6.2 6.5 6.6
Task-number: QTBUG-114953
Change-Id: I11518ccf28ee5a5055ae745a5921837d2f3151b6
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>