qv4codegen: Remove redundant jump instruction for coalesce expressions
For binary expressions with the ?? operator, two cases must be handled
in two code paths. When the lhs is defined, it is used as the result of
the expression. When the lhs is undefined, the rhs expression is used as
a fallback instead. Distinguishing between these two branches can be
done with a test such as CmpEqNull and a conditional jump based on it..

Previously, on top of a conditional JumpTrue to jump to the correct
location based on the test, a second JumpFalse was emitted to perform
the same operation a second time. This is redundant.

The second jump was removed and the logic was slightly reorganized.

Change-Id: I65479752d9fac8c18ef2475b062dd2b5262c372e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>