StyleGenerator: clear config filePath if no image was downloaded
If an item in figma ends up being completely transparent, figma
will skip creating an image URL for it, even if we request one.
And this causes problems, since the BorderImages we use in our
QML templates expect the images to always be present.

Therefore, if we detect that Figma didn't produce a URL for an
image, we simply clear the filePath in the config file.

Becuse of the rather limitied API in Qt for working with
JSON (other than reading), the implementation that clears the
filePath from the config file is a bit less performant that it
theoretically could have been.

Change-Id: Ia06cf3439ab78a151f94da3b54bfdf0f7d90af3c
Reviewed-by: Doris Verria <doris.verria@qt.io>