qv4codegen: Eliminate redundant SetUnwindHandler for try-catch-finally
For try-catch-finally sequences with a catch block, the generator would
start by setting the unwind handler to code from the finally block and
immediately after that set the handler to code from the catch block,
overwriting the last instruction.

In cases with catch blocks, do not emit the first instruction setting
the unwind handler.

Change-Id: I588433c1ad2521b1750f77f2084315a707577a8a
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>