Remove QQuickPixmap argument from QQuickPixmapData ctors
We don't use it anymore.

QQuickPixmapData is the PIMPL behind QQuickPixmap, and e.g.
QObjectPrivate has a Q pointer, so to follow the pattern,
QQuickPixmapData should too. But QQuickPixmapData is reference-counted
and can act as the private data for multiple QQuickPixmap instances,
in case the same image is shown multiple places in a UI. Since it lives
independently, and we never needed the QIntrusiveList of Q-pointers for
anything more than leak prevention, there's no need to pass it to the
ctor either.

Change-Id: I096b0b7cf0c19821b95e30dc99d07665b82a811e
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>