Test QQuickPixmap::loadImageFromDevice()
SlowProvider is slow like QPdfIOHandler is; it's a QQuickImageProvider
rather than a QImageIOHandler for ease of testing, and we invoke it via
a QQuickImage subclass that calls QQuickPixmap::loadImageFromDevice(),
like QQuickPdfPageImage does.

Also get the old dataLeak() test running, as a drive-by. It still isn't
useful in CI though, because it has no built-in way of detecting leaks.

Task-number: QTBUG-114953
Change-Id: I9e2950fbaf4ea69969b3bd10a0d8e624f0e4e8c1
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>