qmldomastcreator: be extra cautious during dom construction in 6.6
Add extra safeguards in 6.6 to avoid qmlls to crash on unimplemented JS
code structures: the scriptNodeStack might contain unexpected items from
unimplemented structures, and takeVariant() or takeList() will at best
return wrong values and at worst trigger an assert.

Instead of asserting and crashing the entire language server, always
check if the current script node is a list or variant before actually
getting it. If there is a mismatch, silently deactivate the
scriptelement support and features such as go to definition to
allow the rest of the language server to continue running (e.g. qmllint
warnings in the editor for example does not need the scriptelement
support).

This commit is not needed in dev, as arguably most of the crashes
happen on JS structures that already exist in dev. Refer to
QTBUG-116392 for crashes on qmlls in dev because of unimplemented
structures.

Task-number: QTBUG-116392
Change-Id: If6625bac45aed469c16f02c9d5abb1fab7521408
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>