StyleGenerator: add usage text to the GUI
To make it even easier to use a generated style, add a section
to the gui that explains how to use the style, which also
provides an easy way to copy the import path

For this to work, we also need to make sure that we store the
full path to the style in the bridge, even when the target
directory was specified from the command line.

A few minor drive-bys: change the folder button to be ToolButton
rather than a flat Button to make it smaller, ensure we use
Label rather than Text to get the proper style font and color,
and ensure that we clear the log whenever we start a new generate.

Change-Id: I9d5294d2f4116e45dfec0eb01e821eaa8c88c4c3
Reviewed-by: Doris Verria <doris.verria@qt.io>