QQmlSA: Add initial fixit section
FixSuggestion unfortunately still lacks class documentation, and we
probably need to rework the API before moving the module out of TP.

However, we should still show the basic usage of the module, including
fixit suggestions. Consequently, add a small example on how to write
FixSuggestions, and add a warning note that they should only be marked
as auto-applicable if the resulting code is still valid.

Pick-to: 6.6
Change-Id: I260021163e3afe2cda97fd82a9bf4e9c26a910e2
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>