StyleGenerator: ensure we remove old QML files before we copy new ones
The current implementation would claim to to copy QML files
from the stylegenerator and into the style folder. But in
reality it sometimes failed, since QFile::copy() will refuse
to overwrite existing files.

This patch will fix this oversight by removings the existing file
before taking the copy. In addition, we'll now also check
that the commands we issue during this process actually succseeds
and otherwise print a warning.

Change-Id: I445ecff8da339fd36bb5465e9b853af88fa41648
Reviewed-by: Doris Verria <doris.verria@qt.io>