QQmlSignalNames: Optimize a bit
We don't have to repeat the strlen("on") and strlen("Changed") because
we can phrase them as constexpr statics. The same holds for the actual
strings. We can store them as latin-1 so that we don't have to construct
temporary QStrings. When calculating the signal name from the handler
name, we don't have to check the signal name length twice. And finally,
we can reserve() the QString to return when creating a handler name.

Change-Id: Ied0c33638d9e72df6360dd04b3f517d72beca21b
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>