qml: fix parser grammar for argument deconstruction
Previously, it was hardcoding a StringLiteralPropertyName for
a BindingIdentifier. This does not make sense because
BindingIdentifier cannot be string literals, they have to be
identifiers.

Also, this is problematic for deconstruction, e.g.:
```
function f(something) {
    let { x = 'x' } = something;  // (1)
    return x;
}
```
In (1), you cannot differentiate between the property
name x and the string literal x, because the parser creates a
string literal with value 'x' for both.
Notably, qmlls cannot differentiate between a 'find usages' on the
first x that is supposed to find the x in the return statement, and
the second "x" that is supposed to find no usages because it is a string
literal.

Same happens also for deconstruction in method parameters.

Add tests for qmlls for deconstruction in let-statements, and method
parameters. Make sure qmlformat ignores this qmlfile as it cannot
format deconstruction yet.

Task-number: QTBUG-100084
Change-Id: Ib7df974437e442dc0bfca75381c773c4d30b11bb
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>