qmlls: find usages of method parameters
Fix finding usages and definitions for method parameters: their
definition is not in one of their parents, but in the scope of the
method they belong to.

The findJSIdentifierDefinition() now looks for method parameter
definitions, when no other definition was found, and findUsagesHelper()
tries to add the method parameter definition as usage when it is
missing.

Remove two test cases from tst_qmlls_modules.cpp as they already exist
in tst_qmlls_utils.cpp, and fix those in tst_qmlls_utils as asking for
the definition of a method parameter now returns something. Previously,
it was only returning something when asking for the definition of a
method parameter from a usage inside the method block.

Finally, add a test to find usages of method parameters.

Task-number: QTBUG-100084
Change-Id: I63f85ef41a10509ec6cb39e1afc838e6209d3396
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>