Remove resize guards from qml tool
Resize guards were prohibiting the initialization of the top level
item size to the window size in a qml script when executed with
the qml tool. This is usually fine because the window size is set
to the item size. However, if this is not possible (minimum window
size from the system, no explicit size on the item), the item has
a different size than the window. The item would only be adapted to
the window size after resizing the window.

This patch removes the resize guards to make the item always fill
the window on startup. The resize guard should hopefully not be needed;
the QQuickItem setters for width and height return early if it's
unchanged; QWindow::resize() does not emit changed signals if there
is not yet a platform window and width or height is unchanged;
and we hope all the overrides of QPlatformWindow::setGeometry()
will avoid making unnecessary changes. However the guards were added
in 8d9a7e47aaa03efe2d3eddfc9ac37d02a99affc7 and I don't remember
exactly why.

Adds the test tst_qml::itemAndWindowGeometry to check combinations of
--config, --qwindowgeometry and qml files whose root item has or doesn't
have its own dimensions, or changes its dimensions after being shown.

Done-with: Matthias Rauter <matthias.rauter@qt.io>
Pick-to: 6.5
Fixes: QTBUG-114068
Fixes: QTBUG-116753
Change-Id: Ib972e0bfc25809441d378c53dabb60653314f5a6
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 05e0dc2a88e751e97573610fd483f79da3e78854)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>