Loader: re-set ItemObservesViewport flag on child if already set
When QQuickLoaderPrivate::load() loads a Text item, its ctor calls
QQuickTextPrivate::init(), which calls setFlag(ItemObservesViewport).
QQuickItem::setFlag() tries to go up the parent hierarchy to get
subtreeTransformChangedEnabled turned on; but at that time, the Text
does not yet have a parent.

If a Flickable contains a Loader that loads a Text with large content,
QQuickTextPrivate::transformChanged() needs to get called during
scrolling. Loader is not interested in these notifications for itself;
so its parent's QQuickItemPrivate::transformChanged() detects that and
turns off subtreeTransformChangedEnabled. After the child is loaded,
QQuickLoader::itemChange() detects that the child's ItemObservesViewport
is already set, and sets the same flag again. setFlag() doesn't have a
guard, so the parent traversal happens again, and
subtreeTransformChangedEnabled gets turned back on.

If a different child were loaded, and that child did not want the
notifications, the next QQuickItemPrivate::transformChanged()
would detect it. If the child changed its own flag later on (e.g.
because a small amount of text was replaced with a larger document,
exceeding QQUICKTEXT_LARGETEXT_THRESHOLD), QQuickItem::setFlag() would
be able to go up the parent chain at that time.

For the autotest, long.qml has about 15KB of text. Maybe it will
be reusable. At least it should compress well in git packfiles,
since we have this license text elsewhere.

Fixes: QTBUG-115687
Pick-to: 6.5 6.6
Change-Id: I87b6a42f5735b8f9c2267f91a4112681da05de5d
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>