qmldom: set preCode correctly
MethodParameter name can be potentially empty. In addition, if there
is a leading rest operator or a default assignment, those will not be
seen in preCode. We keep all that information in value. Having correct
preCode is necessary for detecting AST node for the method body.

Amends 599c79569d35fd317e605438f20c1e850060624b

Pick-to: 6.6 6.5
Change-Id: Iacf5be71c0b05018705f5eb6024acac2a6086476
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>