qmlformat: Don't add invalid default settings options
When determining whether or not a qmlformat setting is set in the config
file, QQmlToolingSettings::isSet is called. Having a default value for
a setting is optional, so some settings have an invalid QVariant in the
map already. This means the isSet call always returns true, which
overrides a bunch of command line options.

Now only valid QVariants are added to the initial default settings map,
which prevents command line options from inadvertently being ignored
even if they weren't even set in a config file.

Pick-to: 6.5 6.6
Change-Id: Iea6968a9ced17970ff0b2d79907de32f41b235c2
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>