StyleGenerator: start in GUI mode if no arguments are set
If there are no arguments on the command line, we assume GUI
mode, otherwise command line mode. That way you can first fill
in the figma fields using GUI mode, but later just rerun the
generator from the command line (and using -g can then be
used to limit which controls to generate).

Change-Id: I28132ff81fb4efc70a56163dc83d63a187f42c8e
Reviewed-by: Doris Verria <doris.verria@qt.io>