qmlls: fix DomItem::filterUp and DomItem::rootQmlObject
Rewrite broken DomItem::filterUp method: the code was unreadable
and the search stopped at owner() instead of top(), which makes it
problematic for scriptexpressions when going from a scriptexpression to
the containing QmlComponent, for example, as it stops at owner() before
actually reaching the QmlComponent. Write a testcase for that.
Also add some internal documentation for filterUp and FilterUpOptions,
as Inner and Outer have special meanings.

Also remove coded-duplication in the implementation of rootQmlObject and
qmlObject.

This prepares the DOM for the extension of the existing autocompletion
feature.

Task-number: QTBUG-116899
Change-Id: I3d8c619af0a9eed74e49b0fc81f2c20ad4947ab1
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>