StyleGenerator: improve BorderImage offset calculations
We used to only look at the radii for the "fillAndStroke"
item to determine the border image offsets. This was not
good enough, since the rectangle could also be square with
a thick border. And also in this case do we want to set an
offset to not scale the border itself.

This patch will therefore change it so that we look at
both the radii and the border width to guess the border
offset. Whichever is greatest (per corner/side) wins.

Moreover, since this this strategy sometimes fails if the
background contain additional elements, such as extra
lines or icons on top of the background, we add a new
"borderImageOffset" item inside the primitives. This item
has a layout that lets the designer set an explicit padding
that controls the offset. If this "borderImageOffset" is
visible, we always use i padding. Otherwise, we fall
back to guess the offset based on the "fillAndStroke" item.

Change-Id: Ia07492018e811c20b586d2d879927fcdbb927748
Reviewed-by: Doris Verria <doris.verria@qt.io>