Make sure to listen for itemDestroyed when using QQuickDeferredPointer
When using QQuickDeferredPointer for deferred properties in Qt Quick
Controls, it is necessary to also implement the itemDestroyed virtual
method because if the item pointed gets deleted the
QQuickDeferredPointer reference will be a dangling pointer when cleanup
occurs in the destructor. This patch is following the pattern laid out
by QQuickControlsPrivate for handling its QQuickDeferredPointer members.

This bug is exposed when aborting a component incubation for types that
did not do this cleanup in the form of a segfault in the destructor of
the offending class.

Fixes: QTBUG-116828
Change-Id: I6a486de74ee38ccdcaff6a31195ee389f51fc93c
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 4e41739c5860befa6d4c5d8601661b1309799465)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d733885962d797ad2880374e014e962e2a64ee7c)
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>