qmlls: Fix codemodel test
This test was referencing a non-existent directory as data directory
which immediately produces an assertion. Make it a regular test rather
than a QQmlDataTest. It doesn't need a data directory.

We learn that this test does not run on the CI.

Change-Id: I92b6c21fea10f3026f351470242bcdf48852c167
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit b3dca709994310aa1769379a55520e0910d775a5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>