qmlformat: Fix object declaration indentation in arrays
Object declarations inside of array patterns did not insert newlines
correctly. This fixes those while not touching other types of array
initializations (such as numeric literals) and the one of test data.

When extra commas are added to the end of arrays, they are now kept on the
correct line as well.

Pick-to: 6.5 6.6 6.7
Change-Id: I8fe67ef066b84f56237449c7695990daa915b4b6
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>