qmlls: singletons, attached types and enums completion preparation
Modify and change QQmlLSUtils::completions() to allow adding
completion for more QML and JavaScript language constructs. Add some
tests for autocompletion of singletons, attached types and enums and
QEXPECT_FAIL them: those will be tended in a separate commit.

Fix completions() to find the first parent with the
interesting domtype, to avoid triggering the 'DomType::Binding' codepath
when in fact the current item is inside a QmlObject inside the Binding
and therefore should get the QmlObject-specific completion, for example
in
```
Item {
    property var xxx: Item {/* complete for QmlObject, not Binding
    here*/ }
}
```

Also suppress completion inside of method name definition: the only
completions inside a function definition are the one inside the
function block and the one for type annotations. Previously, qmlls
was generating some already existing names for the autocompletion of the
name of a method definition (which would lead to methods having the same
name as other existing methods/properties/etc, for example).

Try to do the same also for property definitions to avoid autocompleting the
name of a property definition: the only completions inside a property
definition are for the type and the binding (the latter is handled in
the same code that takes care of 'normal' bindings).
This is currently not implemented yet, so mark the added tests for this
as QEXPECT_FAIL.

Task-number: QTBUG-116899
Change-Id: If7bd51c549cc35e64eba6ceafa313d3df403b1ba
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>