Add missing connect context object in QQuickPaletteProviderPrivateBase
When building QtMultimedia on Windows as static library, the build
failed when compiling qtmultimedia's multimediaquick module in

    qtmultimedia\src\multimediaquick\qquickvideooutput.cpp

because a connect statement was contextless. This is no longer
allowed because QT_NO_CONTEXTLESS_CONNECT is defined.

The build issue is fixed by adding the output from itemWithPalette() as
a context object. The output from itemWithPalette() is the object being
used inside the lambda function.

Change-Id: I0d9c807e360ffccf51da46913cd9faa5c71f5705
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>