QML: Silence warnings about converting holes in sparse arrays
Holes in sparse arrays are undefined when read via get(). QV4::Sequence
does not have holes, by design. Therefore, when converting, we fill the
holes with default-constructed values. This is on purpose and should not
cause warnings.

Furthermore, since QVariant::convert() always produces the given type,
we don't need to re-initialize the variant afterwards.

Change-Id: I46a675dfb71ae9f66858c97a580ec133aabef10e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>