CMake: Fix deployment of QML plugins for debug-and-release Qt
For simplicity, whenever we mention debug-and-release in this commit
message, we mean the more general concept of "multi-configuration build
with a debug configuration".

When deploying a QtQuick application that was built against a
debug-and-release Qt (like the Windows one that comes with the
installer), we copied both builds of a QML plugin to the installation
directory.

Now, when we've detected that the used Qt is a debug-and-release build,
we match exactly against the debug suffix if our application is a debug
build, or the absence of the debug suffix if our application is a
release build.

For non-debug-and-release Qt builds we leniently match against any debug
suffix, including the empty one, as before.

Fixes: QTBUG-109444
Change-Id: I0d02cce2be871286fa6e9bac521d692e96c1bbf0
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 854ee1c03c662ba5c059ad3d3a19d2c2dc750d15)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d0513abe1f88c48ea660fe13ec7466a596b8fc33)