Fix Material dark theme not being respected when used as a fallback
The fallback theme was previously not being initialized if the style
plugin for which registerTypes was called was not the current style;
that is, only the current style would have its theme initialized.
As the Material (and Universal) style's settings such as light/dark
theme are taken into use in its style plugin's initializeTheme
function, we must call that.

The fallback style's theme is now initialized before the current style's
theme, meaning that any initialization custom styles' themes may do
in the future (if QTBUG-67062 is implemented and the theming API
becomes public) will overwrite any shared font/palette values,
as expected.

[ChangeLog][Controls][Important Behavior Changes] The fallback style's
theme is now initialized, allowing e.g. the Material style's dark
theme setting to be respected when it's set as the fallback style and
the user is using run-time style selection.
Compile-time style selection is unaffected and continues to work
correctly.

Task-number: QTBUG-86355
Task-number: QTBUG-117403
Change-Id: Ic651816592c8ca828cfc245be738dac20d89be46
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>